move all logic from `App` to `HomePage` page
